Move SelectedTextProvider so we can use it in the GitHub.VisualStudio project.
