+++ src/GitHub.Exports.Reactive/GitHub.Exports.Reactive.csproj
@@ -91,7 +91,6 @@
     <Compile Include="Models\IConnectionRepositoryHostMap.cs" />
     <Compile Include="Models\IRepositoryHosts.cs" />
     <Compile Include="Services\IModelService.cs" />
-    <Compile Include="Services\ISelectedTextProvider.cs" />
     <Compile Include="ViewModels\IGistCreationViewModel.cs" />
     <Compile Include="ViewModels\ILoginControlViewModel.cs" />
     <Compile Include="ViewModels\ILoginToGitHubForEnterpriseViewModel.cs" />
