+++ src/GitHub.Exports/GitHub.Exports.csproj
@@ -112,6 +112,7 @@
     <Compile Include="Services\Connection.cs" />
     <Compile Include="Services\GitService.cs" />
     <Compile Include="Services\IGitService.cs" />
+    <Compile Include="Services\ISelectedTextProvider.cs" />
     <Compile Include="Services\ITeamExplorerServiceHolder.cs" />
     <Compile Include="Services\Logger.cs" />
     <Compile Include="Services\Services.cs" />
